V2.1.e
mqtt function

+++ ArduinoLibrary/libraries/ArduinoJson/extras/fuzzing/fuzzer_main.cpp
@@ -1,5 +1,5 @@
 // ArduinoJson - arduinojson.org
-// Copyright Benoit Blanchon 2014-2019
+// Copyright Benoit Blanchon 2014-2020
 // MIT License
 
 // This file is NOT use by Google's OSS fuzz
@@ -21,7 +21,7 @@ std::vector<uint8_t> read(const char* path) {
   }
 
   fseek(f, 0, SEEK_END);
-  size_t size = ftell(f);
+  size_t size = static_cast<size_t>(ftell(f));
   fseek(f, 0, SEEK_SET);
 
   std::vector<uint8_t> buffer(size);
